Work around Migen's finalization being subtly order dependent.
Right now, in Migen, if the module tree looks like this:

    top
    \---- probe
    \---- microscope

and finalizing `microscope` makes the module tree look like this:

    top
    \---- probe ----- ps_arm
    \---- microscope

whether any Verilog will actually be generated for `ps_arm` is
dependent on traversal order of submodules, which cannot be
relied upon. This causes `microscope /dev/ttyX singles` to hang.

Avoid this by making finalization of probes order-independent.